+++ R/stat_contour2.r
@@ -50,6 +50,12 @@ StatContour2 <- ggplot2::ggproto("StatContour2", Stat,
   compute_group = function(data, scales, bins = NULL, binwidth = NULL,
                            breaks = scales::fullseq, complete = FALSE,
                            na.rm = FALSE, circular = NULL) {
+      z <- tapply(data$z, data[c("x", "y")], identity)
+
+      if (is.list(z)) {
+          stop("Contour requires single `z` at each combination of `x` and `y`.",
+               call. = FALSE)
+      }
       # Check is.null(breaks) for backwards compatibility
       if (is.null(breaks)) {
           breaks <- scales::fullseq
@@ -73,7 +79,13 @@ StatContour2 <- ggplot2::ggproto("StatContour2", Stat,
           data <- RepeatCircular(data, circular)
       }
       contours <- as.data.table(.contour_lines(data, breaks, complete = complete))
+
+      if (length(contours) == 0) {
+          warning("Not possible to generate contour data", call. = FALSE)
+          return(data.frame())
+      }
       contours <- .order_contour(contours, setDT(data))
+
       return(contours)
     }
 )